requirements.txt
playwright>=1.40.0
pandas>=2.0.0
plotly>=5.18.0
python-dateutil>=2.8.0
streamlit>=1.30.0
openai>=1.0.0  # For xAI Grok API (OpenAI-compatible)
pyyaml>=6.0
pytest>=7.0
python-dotenv>=1.0.0
resend>=0.8.0
tenacity>=8.0.0  # Retry with exponential backoff for API calls
psycopg2-binary>=2.9.9  # PostgreSQL support (optional, for Neon cloud DB)
